requirements.txt
aiogram==3.13.1
gspread==6.1.0
google-auth==2.34.0
